% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_revenue.R
\name{calc_fish_revenue}
\alias{calc_fish_revenue}
\title{computes most frequently caught fish, state revenue from fisheries, total fisheries revenue & plots revenue per state}
\usage{
calc_fish_revenue(fish_caught, fish_prices, plot)
}
\arguments{
\item{fish_caught}{}

\item{fish_prices}{}

\item{plot}{}
}
\value{
lists of answers and plot
}
\description{
computes most frequently caught fish, state revenue from fisheries, total fisheries revenue & plots revenue per state
}
