% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKgas.R
\docType{data}
\name{UKgas}
\alias{UKgas}
\title{UKgas}
\format{An object of class \code{ts} of length 108.}
\usage{
UKgas
}
\description{
Guessing it's UK gas prices?
}
\details{
\@format
 \@source
\\itemize{
\item Quarter 1
}
}
\keyword{datasets}
